requirements.txt
torch
gymnasium
ale-py
numpy
Pillow
torchvision
imageio
